fixed the department edit and implemented the course edit

+++ School.DAT/Models/CourseDaoModel.cs
@@ -4,7 +4,7 @@
     {
         public int CourseId { get; set; }
         public string? Title { get; set; }
-        public decimal Credits { get; set; }
+        public int Credits { get; set; }
         public int DepartmentId { get; set; }
         public string? DepartmentName { get; set; }
         public DateTime? CreatedDate { get; set; }
